prospek_kerja-scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
tqdm==4.67.1
openpyxl==3.1.5
google-genai==1.0.0
